+++ utils/stringHelpers.js
@@ -1,0 +1,5 @@
+function normalizarServico(servicoNome) {
+  return servicoNome.toLowerCase().replace(/\s+/g, '');
+}
+
+module.exports = { normalizarServico };
